% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acctbly.R
\docType{data}
\name{acctbly_dist}
\alias{acctbly_dist}
\title{District Accountability Data for Kentucky Public Schools, 2012-2015}
\format{A dataframe with 894 rows and 8 variables:
\describe{
  \item{sch_id}{ID number to identify schools & districts}
  \item{dist_name}{District name}
  \item{year}{School year}
  \item{acctbly_score}{Total accountabiltiy score}
  \item{acctbly_label}{Accountability label}
  \item{reward_assist}{Reward/assistance label}
  \item{prog_label}{Indicates "Progressing" addition to accountabiltiy label}
  \item{high_progress}{Indicates high progress}
}}
\source{
\url{https://applications.education.ky.gov/src/DataSets.aspx}
}
\usage{
acctbly_dist
}
\description{
A dataset with accountabiltiy data
}
\keyword{datasets}
