% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acctbly.R
\docType{data}
\name{acctbly_essa_sch}
\alias{acctbly_essa_sch}
\title{School Accountability Data for Kentucky Public Schools, 2017-18}
\format{A dataframe with 1272 rows and 14 variables:
\describe{
  \item{sch_id}{ID number to identify schools & districts}
  \item{dist_name}{District name}
  \item{sch_name}{School name}
  \item{year}{School year}
  \item{title1}{Title I status}
  \item{school_level}{School level (Elementary, Middle, or High School)}
  \item{acctbly_label}{Accountability label}
  \item{acctbly_label_reason}{Reason or reasons for accountability label}
  \item{prof_indicator}{Proficiency indicator for accountability. Includes reading and mathematics.}
  \item{sep_acad_indicator}{Separate academic indicator for accountability. Includes writing, social studies, and science.}
  \item{growth_indicator}{Student growth indicator for accountability. Based on growth in reading and mathematics.}
  \item{transition_indicator}{Transition readiness indicator for accountability}
  \item{grad_indicator}{Graduation indicator for accountability}
  \item{grad_rate}{4-year cohort graduation rate}
}}
\source{
\url{http://openhouse.education.ky.gov}
}
\usage{
acctbly_essa_sch
}
\description{
A dataset with accountabiltiy data under the reauthorized ESSA
}
\keyword{datasets}
