% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acctbly.R
\docType{data}
\name{acctbly_state}
\alias{acctbly_state}
\title{State Accountability Data for Kentucky Public Schools, 2012-2015}
\format{A dataframe with 5 rows and 6 variables:
\describe{
  \item{sch_id}{ID number to identify schools & districts}
  \item{dist_name}{District name - in this case, it is always "State Total"}
  \item{year}{School year}
  \item{acctbly_score}{Total accountabiltiy score}
  \item{acctbly_label}{Accountability label}
  \item{prog_label}{Indicates "Progressing" addition to accountabiltiy label}
}}
\source{
\url{https://applications.education.ky.gov/src/DataSets.aspx}
}
\usage{
acctbly_state
}
\description{
A dataset with accountabiltiy data
}
\keyword{datasets}
