% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ach_grade.R
\docType{data}
\name{ach_grade_sch}
\alias{ach_grade_sch}
\title{School KPREP Grade Level Data for Kentucky Public Schools, 2012-2017}
\format{A dataframe with 1,320,672 rows and 15 variables:
\describe{
  \item{sch_id}{ID number to identify schools & districts}
  \item{dist_name}{District name}
  \item{sch_name}{School name}
  \item{year}{School year}
  \item{grade}{Grade level - "EO" indicates high school end-of-course exam that may be administered to students in various grades}
  \item{subject}{Specific subject assessed by KPREP}
  \item{student_group}{Student subgroup}
  \item{n_tested}{Number of students tested}
  \item{novice_pct}{Percent of students scoring "Novice" on KPREP}
  \item{apprentice_pct}{Percent of students scoring "Apprentice" on KPREP}
  \item{proficient_pct}{Percent of students scoring "Proficient" on KPREP}
  \item{distinguished_pct}{Percent of students scoring "Distinguished" on KPREP}
  \item{prof_dist_pct}{Percent of students scoring "Proficient" or "Distinguished" on KPREP}
  \item{bonus_pct}{Percent difference between "Distinguished" and "Novice" if the "Novice" rate is lower.}
  \item{napd_calc}{Aggregation of score granting .5 points per student at "Apprentice", 1 point per student at/above "Proficient" and .5 points for the value of the "Bonus" amount.}
  \item{perf_index}{Aggregation of performance levels, granting weights of 0 to percent "Novice", .5 to percent "Apprentice", 1 to percent "Proficient", and 1.25 to percent "Distinguished".}
}}
\source{
\url{https://applications.education.ky.gov/src/DataSets.aspx}
}
\usage{
ach_grade_sch
}
\description{
A dataset with KPREP performance level data by grade
}
\keyword{datasets}
