% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safety.R
\docType{data}
\name{behavior_events_sch}
\alias{behavior_events_sch}
\title{School Behavior Event Data for Kentucky Public Schools, 2012-2017}
\format{A dataframe with 816,277 rows and 8 variables:
\describe{
  \item{sch_id}{ID number to identify schools & districts}
  \item{dist_name}{District name}
  \item{sch_name}{School name}
  \item{year}{School year}
  \item{category}{Behavior event type, total behavior events, or violations per 1,000 ADA}
  \item{group_type}{Student group type, either race, gender, or total}
  \item{student_group}{Student group}
  \item{n_students}{Number of students}
}}
\source{
\url{https://applications.education.ky.gov/src/DataSets.aspx}
}
\usage{
behavior_events_sch
}
\description{
A dataset with behavior event data at the school level
}
\keyword{datasets}
