% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safety.R
\docType{data}
\name{behavior_ses_state}
\alias{behavior_ses_state}
\title{State Behavior Event Socioeconomic Data for Kentucky Public Schools, 2012-2017}
\format{A dataframe with 198 rows and 7 variables:
\describe{
  \item{sch_id}{ID number to identify schools & districts}
  \item{dist_name}{District name}
  \item{year}{School year}
  \item{category}{Student lunch program status}
  \item{group_type}{Student group type, either race, gender, or total}
  \item{student_group}{Student group}
  \item{n_students}{Number of students}
}}
\source{
\url{https://applications.education.ky.gov/src/DataSets.aspx}
}
\usage{
behavior_ses_state
}
\description{
A dataset with behavior event socioeconomic data at the state level
}
\keyword{datasets}
