% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccr.R
\docType{data}
\name{ccr_state}
\alias{ccr_state}
\title{State College/Career Readiness Data for Kentucky Public Schools, 2012-2016}
\format{A dataframe with 85 rows and 9 variables:
\describe{
  \item{sch_id}{ID number to identify schools & districts}
  \item{dist_name}{District name - in this case, it is always "State Total"}
  \item{year}{School year}
  \item{student_group}{Student subgroup}
  \item{ccr_pct}{Percent of students college and/or career ready}
  \item{college_only_pct}{Percent of students college ready but not career ready}
  \item{career_only_pct}{Percent of students career ready but not college ready}
  \item{both_college_career_pct}{Percent of students both  college and career ready}
  \item{no_ccr_pct}{Percent of students neither college ready or career ready}
}}
\source{
\url{https://applications.education.ky.gov/src/DataSets.aspx}
}
\usage{
ccr_state
}
\description{
A dataset with college/career ready data
}
\keyword{datasets}

