% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safety.R
\docType{data}
\name{discipline_sch}
\alias{discipline_sch}
\title{School Discipline Data for Kentucky Public Schools, 2012-2017}
\format{A dataframe with 731,918 rows and 8 variables:
\describe{
  \item{sch_id}{ID number to identify schools & districts}
  \item{dist_name}{District name}
  \item{sch_name}{School name}
  \item{year}{School year}
  \item{category}{Discipline type}
  \item{group_type}{Student group type, either race, gender, or total}
  \item{student_group}{Student group}
  \item{n_students}{Number of students}
}}
\source{
\url{https://applications.education.ky.gov/src/DataSets.aspx}
}
\usage{
discipline_sch
}
\description{
A dataset with discipline data at the school level
}
\keyword{datasets}
