% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/race.R
\docType{data}
\name{dist_race_pct}
\alias{dist_race_pct}
\title{District Student Race Data for Kentucky Public Schools, 2012-2015}
\format{A dataframe with 28 rows and 5 variables:
\describe{
  \item{sch_id}{ID number to identify schools & districts}
  \item{dist_name}{District name}
  \item{year}{School year}
  \item{race}{Racial subgroup}
  \item{pct}{Percent of students enrolled of a particular race}
}}
\source{
\url{https://applications.education.ky.gov/src/DataSets.aspx}
}
\usage{
dist_race_pct
}
\description{
A dataset with student racial group data
}
\keyword{datasets}

