% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teach.R
\docType{data}
\name{dist_teach_gender_count}
\alias{dist_teach_gender_count}
\title{District Teacher Gender Count Data for Kentucky Public Schools, 2013-2015}
\format{A dataframe with 1040 rows and 5 variables:
\describe{
  \item{sch_id}{ID number to identify schools & districts}
  \item{dist_name}{District name}
  \item{year}{School year}
  \item{t_gender}{Teacher gender}
  \item{count}{Number of teachers}
}}
\source{
\url{https://applications.education.ky.gov/src/DataSets.aspx}
}
\usage{
dist_teach_gender_count
}
\description{
A dataset with teacher gender count data
}
\keyword{datasets}

