% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finance.R
\docType{data}
\name{fin_summary_state}
\alias{fin_summary_state}
\title{State Financial Summary Data for Kentucky Public Schools, 2012-2016}
\format{A dataframe with 65 rows and 6 variables:
\describe{
  \item{sch_id}{ID number to identify schools & districts}
  \item{dist_name}{District name - in this case, it is always "State Total"}
  \item{year}{School year}
  \item{type}{Type of data, either `summary` or `salary`}
  \item{category}{Variable name, which includes the following: `ada` - average daily attendance; `ada_seek` - average daily attendance used for SEEK formula calculations, `adm` - average daily membership; `fte_classified` - full time equivilent classified employees; `fte_staff` - full time equivilent staff; `fte_teachers` - full time equivilent teachers; `fund_bal_pct` - percent of general fund remaining; `fund_balance` - general fund balance; `salary_bene_pct` - percent of expenditures going to salaries and benefits, `avg_p_sal` - average principal salary; `avg_t_sal` - average teacher salary; `sub_exp` - expenditures on substitute teachers, `sub_exp_pp` - per pupil expentitures on substitute teachers}
  \item{value}{Numerical value of variable in `category` column}
}}
\source{
\url{https://applications.education.ky.gov/src/DataSets.aspx}
}
\usage{
fin_summary_state
}
\description{
A dataset with financial summary data
}
\keyword{datasets}
