% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frpl.R
\docType{data}
\name{frpl_sch}
\alias{frpl_sch}
\title{School FRPL Data for Kentucky Public Schools, 2012-2016}
\format{A dataframe with 7159 rows and 7 variables:
\describe{
  \item{sch_id}{ID number to identify schools & districts}
  \item{dist_name}{District name}
  \item{sch_name}{School name}
  \item{year}{School year}
  \item{enroll}{Total student enrollment}
  \item{frpl_total}{Number of students qualifying for free or reduced-price lunch}
  \item{frpl_pct}{Percent of students qualifying for free or reduced-price lunch}
}}
\source{
\url{https://applications.education.ky.gov/src/DataSets.aspx}
}
\usage{
frpl_sch
}
\description{
A dataset with student FRPL data
}
\keyword{datasets}
