% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frpl.R
\docType{data}
\name{frpl_state}
\alias{frpl_state}
\title{State FRPL Data for Kentucky Public Schools, 2012-2016}
\format{A dataframe with 5 rows and 6 variables:
\describe{
  \item{sch_id}{ID number to identify schools & districts}
  \item{dist_name}{District name - in this case, it is always "State Total"}
  \item{year}{School year}
  \item{enroll}{Total student enrollment}
  \item{frpl_total}{Number of students qualifying for free or reduced-price lunch}
  \item{frpl_pct}{Percent of students qualifying for free or reduced-price lunch}
}}
\source{
\url{https://applications.education.ky.gov/src/DataSets.aspx}
}
\usage{
frpl_state
}
\description{
A dataset with student FRPL data
}
\keyword{datasets}
