% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grad.R
\docType{data}
\name{grad_sch}
\alias{grad_sch}
\title{School 4-Year Cohort Graduation Rate Data for Kentucky Public Schools, 2013-2016}
\format{A dataframe with 13680 rows and 6 variables:
\describe{
  \item{sch_id}{ID number to identify schools & districts}
  \item{dist_name}{District name}
  \item{sch_name}{School name}
  \item{student_group}{Student subgroup}
  \item{year}{School year}
  \item{grad_rate}{4-year cohort graduation rate}
}}
\source{
\url{https://applications.education.ky.gov/src/DataSets.aspx}
}
\usage{
grad_sch
}
\description{
A dataset with graduation rate data
}
\keyword{datasets}

