% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grad.R
\docType{data}
\name{grad_state}
\alias{grad_state}
\title{State 4-Year Cohort Graduation Rate Data for Kentucky Public Schools, 2013-2017}
\format{A dataframe with 75 rows and 5 variables:
\describe{
  \item{sch_id}{ID number to identify schools & districts}
  \item{dist_name}{District name - in this case, it is always "State Total"}
  \item{student_group}{Student subgroup}
  \item{year}{School year}
  \item{grad_rate}{4-year cohort graduation rate}
}}
\source{
\url{https://applications.education.ky.gov/src/DataSets.aspx}
}
\usage{
grad_state
}
\description{
A dataset with graduation rate data
}
\keyword{datasets}
