% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iep.R
\docType{data}
\name{iep_sch}
\alias{iep_sch}
\title{IEP Data for Kentucky Public Schools, 2012-2017}
\format{A dataframe with 161,712 rows and 7 variables:
\describe{
  \item{sch_id}{ID number to identify schools & districts}
  \item{dist_name}{District name}
  \item{sch_name}{School name}
  \item{year}{School year}
  \item{student_group}{Student group}
  \item{iep_total}{Total number of students with an IEP}
  \item{iep_pct}{IEP students as percentage of school population}
}}
\source{
\url{https://applications.education.ky.gov/src/DataSets.aspx}
}
\usage{
iep_sch
}
\description{
A dataset with school-level data on students with Individualized Education Plans (IEP)
}
\keyword{datasets}
