% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prog.R
\docType{data}
\name{iep_state}
\alias{iep_state}
\title{Data on students with IEP's in Kentucky Public Schools, 2013-2016}
\format{A dataframe with 96 rows and 6 variables:
\describe{
  \item{sch_id}{ID number to identify schools & districts}
  \item{dist_name}{District name - in this case, it is always "State"}
  \item{year}{School year}
  \item{student_group}{Student subgroup}
  \item{iep_total}{Students with IEP's}
  \item{iep_pct}{Percent of students with IEP's}
}}
\source{
\url{https://applications.education.ky.gov/src/DataSets.aspx}
}
\usage{
iep_state
}
\description{
A dataset with data on students with Individualized Educaitonal Plans
}
\keyword{datasets}

