% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safety.R
\docType{data}
\name{legal_dist}
\alias{legal_dist}
\title{District Legal Action Data for Kentucky Public Schools, 2012-2017}
\format{A dataframe with 58,014 rows and 7 variables:
\describe{
  \item{sch_id}{ID number to identify schools & districts}
  \item{dist_name}{District name}
  \item{year}{School year}
  \item{category}{Legal action type}
  \item{group_type}{Student group type, either race, gender, or total}
  \item{student_group}{Student group}
  \item{n_students}{Number of students}
}}
\source{
\url{https://applications.education.ky.gov/src/DataSets.aspx}
}
\usage{
legal_dist
}
\description{
A dataset with legal action data at the district level
}
\keyword{datasets}
