% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safety.R
\docType{data}
\name{legal_sch}
\alias{legal_sch}
\alias{legal_sch}
\title{School Legal Action Data for Kentucky Public Schools, 2012-2017}
\format{A dataframe with 446,160 rows and 8 variables:
\describe{
  \item{sch_id}{ID number to identify schools & districts}
  \item{dist_name}{District name}
  \item{sch_name}{School name}
  \item{year}{School year}
  \item{category}{Legal action type}
  \item{group_type}{Student group type, either race, gender, or total}
  \item{student_group}{Student group}
  \item{n_students}{Number of students}
}}
\source{
\url{https://applications.education.ky.gov/src/DataSets.aspx}

\url{https://applications.education.ky.gov/src/DataSets.aspx}
}
\usage{
legal_sch

legal_sch
}
\description{
A dataset with legal action data at the school level

A dataset with legal action data at the school level
}
\keyword{datasets}
