% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prog.R
\docType{data}
\name{lep_dist}
\alias{lep_dist}
\title{Limited English Proficiency Data for Kentucky Public Schools, 2013-2017}
\format{A dataframe with 11,115 rows and 6 variables:
\describe{
  \item{sch_id}{ID number to identify schools & districts}
  \item{dist_name}{District name}
  \item{year}{School year}
  \item{student_group}{Student subgroup}
  \item{lep_total}{Students with Limited English Proficiency}
  \item{lep_pct}{Percent of students with Limited English Proficiency}
}}
\source{
\url{https://applications.education.ky.gov/src/DataSets.aspx}

\url{https://applications.education.ky.gov/src/DataSets.aspx}
}
\usage{
lep_dist

lep_dist
}
\description{
A dataset with data on Limited English Proficiency

A dataset with data on students with Individualized Educational Plans
}
\keyword{datasets}
