% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prog.R
\docType{data}
\name{lep_sch}
\alias{lep_sch}
\alias{lep_sch}
\title{Limited English Proficiency Data for Kentucky Public Schools, 2013-2016}
\format{A dataframe with 57,915 rows and 7 variables:
\describe{
  \item{sch_id}{ID number to identify schools & districts}
  \item{dist_name}{District name}
  \item{sch_name}{School name}
  \item{year}{School year}
  \item{student_group}{Student subgroup}
  \item{lep_total}{Students with Limited English Proficiency}
  \item{lep_pct}{Percent of students with Limited English Proficiency}
}}
\source{
\url{https://applications.education.ky.gov/src/DataSets.aspx}

\url{https://applications.education.ky.gov/src/DataSets.aspx}
}
\usage{
lep_sch

lep_sch
}
\description{
A dataset with data on Limited English Proficiency

A dataset with data on students with Individualized Educational Plans
}
\keyword{datasets}
