% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\docType{data}
\name{profile_dist}
\alias{profile_dist}
\title{District Profile Data for Kentucky Public Schools, 2012-2015}
\format{A dataframe with 704 rows and 8 variables:
\describe{
  \item{sch_id}{ID number to identify schools & districts}
  \item{dist_name}{District name}
  \item{coop}{Regional cooperative}
  \item{long}{Longitude}
  \item{lat}{Latitude}
  \item{year}{A factor containing the school year}
  \item{enroll}{Number of students enrolled on October 1 of that school year}
}}
\source{
\url{https://applications.education.ky.gov/src/DataSets.aspx}
}
\usage{
profile_dist
}
\description{
A dataset with basic location and enrollment data
}
\keyword{datasets}

