% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\docType{data}
\name{profile_state}
\alias{profile_state}
\title{State Profile Data for Kentucky Public Schools, 2012-2015}
\format{A dataframe with 4 rows and 6 variables:
\describe{
  \item{sch_id}{ID number to identify schools & districts}
  \item{dist_name}{District name - in this case, it is always "State Total"}
  \item{long}{Longitude}
  \item{lat}{Latitude}
  \item{year}{School year}
  \item{enroll}{Number of students enrolled on October 1 of that school year}
}}
\source{
\url{https://applications.education.ky.gov/src/DataSets.aspx}
}
\usage{
profile_state
}
\description{
A dataset with basic location and enrollment data
}
\keyword{datasets}
