% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/race.R
\docType{data}
\name{race_state}
\alias{race_state}
\title{State Student Race Count Data for Kentucky Public Schools, 2012-2016}
\format{A dataframe with 35 rows and 6 variables:
\describe{
  \item{sch_id}{ID number to identify schools & districts}
  \item{dist_name}{District name - in this case, it is always "State Total"}
  \item{year}{School year}
  \item{race}{Racial subgroup}
  \item{s_count}{Number of students enrolled of a particular race}
  \item{calc_pct}{Calculated relative percentage of student enrollment}
}}
\source{
\url{https://applications.education.ky.gov/src/DataSets.aspx}
}
\usage{
race_state
}
\description{
A dataset with student racial group count data
}
\keyword{datasets}
