% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finance.R
\docType{data}
\name{rev_exp_dist}
\alias{rev_exp_dist}
\title{District Revenue and Expenditure Data for Kentucky Public Schools, 2012-2016}
\format{A dataframe with 25,143 rows and 7 variables:
\describe{
  \item{sch_id}{ID number to identify schools & districts}
  \item{dist_name}{District name}
  \item{year}{School year}
  \item{type}{Indicates if variable is `revenue` or `expenditure` data}
  \item{tot_pp}{Indicates if varaible is `total` amount or `pp` - per pupil, which is calculated by dividing the total by `adm` - average daily membership}
  \item{category}{Variable name, which includes the following: `admin` - administration; `admin_pct` - percent of expenditures spent on administration; `capital_outlay` - capital outlay; `enterprise_ops` - enterprise operations; `federal` - federal revenue; `instr_staff_support` - instrucitonal staff support; `instruction` - instruciton; `local` - local revenue; `ops_maint` - operations and maintenance; `state` - state revenue; `student_support` - student supports; `total` - total amount; `transport` - transportation}
  \item{value}{Numerical value of variable in `category` column}
}}
\source{
\url{https://applications.education.ky.gov/src/DataSets.aspx}
}
\usage{
rev_exp_dist
}
\description{
A dataset with revenue and expenditure data
}
\keyword{datasets}
