% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\docType{data}
\name{sch_profile}
\alias{sch_profile}
\title{School Profile Data for Kentucky Public Schools, 2012-2015}
\format{A dataframe with 5722 rows and 7 variables:
\describe{
  \item{sch_id}{ID number to identify schools & districts}
  \item{dist_name}{District name}
  \item{sch_name}{School name}
  \item{long}{Longitude}
  \item{lat}{Latitude}
  \item{year}{School year}
  \item{enroll}{Number of students enrolled on October 1 of that school year}
}}
\source{
\url{https://applications.education.ky.gov/src/DataSets.aspx}
}
\usage{
sch_profile
}
\description{
A dataset with basic location and enrollment data
}
\keyword{datasets}

