% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/race.R
\docType{data}
\name{sch_race_count}
\alias{sch_race_count}
\title{School Student Race Count Data for Kentucky Public Schools, 2012-2015}
\format{A dataframe with 38017 rows and 6 variables:
\describe{
  \item{sch_id}{ID number to identify schools & districts}
  \item{dist_name}{District name}
  \item{sch_name}{School name}
  \item{year}{School year}
  \item{race}{Racial subgroup}
  \item{count}{Number of students enrolled of a particular race}
}}
\source{
\url{https://applications.education.ky.gov/src/DataSets.aspx}
}
\usage{
sch_race_count
}
\description{
A dataset with student racial group count data
}
\keyword{datasets}

