% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teach.R
\docType{data}
\name{sch_teach_stats}
\alias{sch_teach_stats}
\title{School Teacher Data for Kentucky Public Schools, 2012-2015}
\format{A dataframe with 5721 rows and 9 variables:
\describe{
  \item{sch_id}{ID number to identify schools & districts}
  \item{dist_name}{District name}
  \item{sch_name}{School name}
  \item{year}{School year}
  \item{s_t_ratio}{Number of students per teacher}
  \item{t_fte}{Full Time Equivelent Teachers}
  \item{nbct_count}{Number of National Board Certified Teachers}
  \item{nbct_pct}{Percent of National Board Certified Teachers}
  \item{avg_t_exp}{Average years of experience}
}}
\source{
\url{https://applications.education.ky.gov/src/DataSets.aspx}
}
\usage{
sch_teach_stats
}
\description{
A dataset with teacher data
}
\keyword{datasets}

