% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finance.R
\docType{data}
\name{seek_dist}
\alias{seek_dist}
\title{District SEEK Data for Kentucky Public Schools, 2012-2017}
\format{A dataframe with 8,320 rows and 5 variables:
\describe{
  \item{sch_id}{ID number to identify schools & districts}
  \item{dist_name}{District name}
  \item{year}{School year}
  \item{category}{Variable name, which includes the following: `actual_transportation_expenditures` - actual transportation expenditures; `building_local` - building fund from local sources; `building_state` - building fund from SEEK; `net_gen_fund` - total SEEK funding less restricted capital outlay; `net_gen_fund_per_pupil` - per pupil SEEK funding less restricted capital outlay; `transp_funded_pct` - SEEK transportation funding divided by total transportation spending; `transportation_seek_funding` - SEEK transportation funding}
  \item{value}{Numerical value of variable in `category` column}
}}
\source{
\url{https://applications.education.ky.gov/src/DataSets.aspx}
}
\usage{
seek_dist
}
\description{
A dataset with revenue and expenditure data
}
\keyword{datasets}
