% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kprep_levels.R
\docType{data}
\name{state_eoc_levels}
\alias{state_eoc_levels}
\title{State KPREP EOC Levels Data for Kentucky Public Schools, 2012-2015}
\format{A dataframe with 288 rows and 11 variables:
\describe{
  \item{sch_id}{ID number to identify schools & districts}
  \item{dist_name}{District name - in this case, it is always "State Total"}
  \item{year}{School year}
  \item{subject}{Specific subject assessed by KPREP EOC exam}
  \item{student_group}{Student subgroup}
  \item{n_tested}{Number of students tested}
  \item{novice_pct}{Percent of students scoring "Novice" on KPREP}
  \item{apprentice_pct}{Percent of students scoring "Apprentice" on KPREP}
  \item{proficient_pct}{Percent of students scoring "Proficient" on KPREP}
  \item{distinguished_pct}{Percent of students scoring "Distinguished" on KPREP}
  \item{prof_dist_pct}{Percent of students scoring "Proficient" or "Distinguished" on KPREP}
}}
\source{
\url{https://applications.education.ky.gov/src/DataSets.aspx}
}
\usage{
state_eoc_levels
}
\description{
A dataset with KPREP performance level data
}
\keyword{datasets}

