% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/race.R
\docType{data}
\name{state_race}
\alias{state_race}
\title{State Student Race Data for Kentucky Public Schools, 2012-2015}
\format{A dataframe with 28 rows and 5 variables:
\describe{
  \item{sch_id}{ID number to identify schools & districts}
  \item{dist_name}{District name - in this case, it is always "State Total"}
  \item{year}{School year}
  \item{race}{Racial subgroup}
  \item{enroll}{Number of students enrolled of a particular race}
}}
\source{
\url{https://applications.education.ky.gov/src/DataSets.aspx}
}
\usage{
state_race
}
\description{
A dataset with student racial group data
}
\keyword{datasets}

