% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teach.R
\docType{data}
\name{state_teach_gender_pct}
\alias{state_teach_gender_pct}
\title{State Teacher Gender Percent Data for Kentucky Public Schools, 2013-2014}
\format{A dataframe with 4 rows and 5 variables:
\describe{
  \item{sch_id}{ID number to identify schools & districts}
  \item{dist_name}{District name - in this case, it is always "State Total"}
  \item{year}{School year}
  \item{t_gender}{Teacher gender}
  \item{pct}{Percent of teachers}
}}
\source{
\url{https://applications.education.ky.gov/src/DataSets.aspx}
}
\usage{
state_teach_gender_pct
}
\description{
A dataset with teacher gender percent data
}
\keyword{datasets}

