% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teach.R
\docType{data}
\name{state_teach_race_pct}
\alias{state_teach_race_pct}
\title{State Teacher Race Percent Data for Kentucky Public Schools, 2014-2015}
\format{A dataframe with 13 rows and 5 variables:
\describe{
  \item{sch_id}{ID number to identify schools & districts}
  \item{dist_name}{District name - in this case, it is always "State Total"}
  \item{year}{School year}
  \item{t_race}{Teacher racial group}
  \item{pct}{Percent of teachers}
}}
\source{
\url{https://applications.education.ky.gov/src/DataSets.aspx}
}
\usage{
state_teach_race_pct
}
\description{
A dataset with teacher race percent data
}
\keyword{datasets}

