% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finance.R
\docType{data}
\name{tax_state}
\alias{tax_state}
\title{State Tax Data for Kentucky Public Schools, 2012-2017}
\format{A dataframe with 24 rows and 5 variables:
\describe{
  \item{sch_id}{ID number to identify schools & districts}
  \item{dist_name}{District name - in this case, it is always "State Total"}
  \item{year}{School year}
  \item{category}{Variable name, which includes the following: `assessment` - value of assessed property including real estate, tangible property, and motor vehicles; `assessment_pp` - per pupil value of assessed property; `property` - tax rate for tangible property in cents per $100; `real_estate` - tax rate for real estate in cents per $100}
  \item{value}{Numerical value of variable in `category` column}
}}
\source{
\url{https://applications.education.ky.gov/src/DataSets.aspx}
}
\usage{
tax_state
}
\description{
A dataset with tax data
}
\keyword{datasets}
