% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccr.R
\docType{data}
\name{ccr_sch}
\alias{ccr_sch}
\title{School College/Career Readiness Data for Kentucky Public Schools, 2012-2016}
\format{A dataframe with 19448 rows and 7 variables:
\describe{
  \item{sch_id}{ID number to identify schools & districts}
  \item{dist_name}{District name}
  \item{sch_name}{School name}
  \item{year}{School year}
  \item{student_group}{Student subgroup}
  \item{ccr_pct}{Percent of students college and/or career ready}
  \item{college_ready_pct}{Percent of students college ready}
}}
\source{
\url{https://applications.education.ky.gov/src/DataSets.aspx}
}
\usage{
ccr_sch
}
\description{
A dataset with college/career ready data
}
\keyword{datasets}

