% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccr.R
\docType{data}
\name{ccr_state}
\alias{ccr_state}
\title{State College/Career Readiness Data for Kentucky Public Schools, 2012-2016}
\format{A dataframe with 85 rows and 11 variables:
\describe{
  \item{sch_id}{ID number to identify schools & districts}
  \item{dist_name}{District name - in this case, it is always "State Total"}
  \item{year}{School year}
  \item{student_group}{Student subgroup}
  \item{n_grads}{Number of students in graduating class}
  \item{college_ready}{Graduating students meeting college-ready benchmarks}
  \item{career_ready}{Graduating students meeting career-ready benchmarks}
  \item{ccr_pct}{Percent of students college and/or career ready}
  \item{college_ready_pct}{Percent of graduating students meeting college-ready benchmarks}
  \item{career_ready_pct}{Percent of graduating students meeting career-ready benchmarks}
}}
\source{
\url{https://applications.education.ky.gov/src/DataSets.aspx}
}
\usage{
ccr_state
}
\description{
A dataset with college/career ready data
}
\keyword{datasets}
