% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iep.R, R/prog.R
\docType{data}
\name{iep_state}
\alias{iep_state}
\alias{iep_state}
\title{IEP Data for Kentucky Public Schools, 2012-2015}
\format{A dataframe with 96 rows and 6 variables:
\describe{
  \item{sch_id}{ID number to identify schools & districts}
  \item{dist_name}{District name - always "State Total"}
  \item{year}{School year}
  \item{student_group}{Student group}
  \item{iep_total}{Total number of students with an IEP}
  \item{iep_pct}{IEP students as percentage of state population}
}}
\source{
\url{https://applications.education.ky.gov/src/DataSets.aspx}

\url{https://applications.education.ky.gov/src/DataSets.aspx}
}
\usage{
iep_state

iep_state
}
\description{
A dataset with state-level data on students with Individualized Education Plans (IEP)

A dataset with data on students with Individualized Educational Plans
}
\keyword{datasets}
